Add authentication to enroller.
Build and push enroller to registry.
